requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
altgraph==0.17.4
asgiref==3.9.1
attrs==25.3.0
babel==2.17.0
bcrypt==5.0.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
contourpy==1.3.3
crispy-bootstrap5==2025.4
cryptography==46.0.3
cssselect2==0.8.0
cycler==0.12.1
Django==5.2.4
django-crispy-forms==2.4
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
fonttools==4.60.1
frozenlist==1.6.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
lxml==6.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
multidict==6.4.3
mysqlclient==2.2.7
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pefile==2023.2.7
phonenumbers==9.0.5
pillow==11.3.0
propcache==0.3.1
pycparser==2.23
pydyf==0.11.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyJWT==2.10.1
pyparsing==3.2.5
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
reportlab==4.4.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tkcalendar==1.6.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
yarl==1.20.0
zopfli==0.2.3.post1
